Migrate User registeration logic to CQRS pattern.
