+++ src/FoodMaster.WebSite/Models/UserDetails.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
+using System.Security.Claims;
+
+using MediatR;
 
 namespace FoodMaster.WebSite.Models
 {
-    public class UserDetails
+    public class UserDetails : IRequest<Claim[]>
     {
         [Required]
         [DataType(DataType.Text)]
